requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
m3u8==5.0.0
multidict==6.0.5
pycryptodome==3.20.0
requests==2.32.3
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
yarl==1.9.4
